feat(transcript): add endpoint to retrieve user transcripts grouped by date

+++ routes/api.php
@@ -14,5 +14,6 @@ Route::delete('/tokens/{id}', [AuthController::class, 'revokeToken'])->middlewar
 
 Route::post('generate-document', [DocumentController::class, 'generate'])->middleware('auth:sanctum');
 Route::get('user/transcripts', [TranscriptController::class, 'indexByUser'])->middleware('auth:sanctum');
-Route::delete('transcripts/{id}', [TranscriptController::class, 'delete']);
-Route::put('transcripts', [TranscriptController::class, 'update']);
+Route::get('user/transcripts/perDate', [TranscriptController::class, 'indexByUserPerDate'])->middleware('auth:sanctum');
+Route::delete('transcripts/{id}', [TranscriptController::class, 'delete'])->middleware('auth:sanctum');
+Route::put('transcripts', [TranscriptController::class, 'update'])->middleware('auth:sanctum');
